requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Research\NVM\ADC\Plugin\model.py: 2
# D:\Research\NVM\ADC\Plugin\optimizer.py: 4
# D:\Research\NVM\ADC\Plugin\run.py: 6
PyYAML == 5.4.1

# D:\Research\NVM\ADC\Plugin\model.py: 6
numpy == 1.20.2

# D:\Research\NVM\ADC\Plugin\model.py: 4
# D:\Research\NVM\ADC\Plugin\murmannsurvey.py: 18
pandas == 1.2.3

# D:\Research\NVM\ADC\Plugin\murmannsurvey.py: 12
regex == 2021.4.4

# D:\Research\NVM\ADC\Plugin\model.py: 5
sklearn == 0.0
